Changed UI files structire
