--- src/popup/style/_font_select.less
@@ -1,109 +1,0 @@
-﻿@import "_variables.less";
-
-/*----------------
-    Font select
-------------------*/
-
-@fontSelectCheckboxMargin: @size-control-inner + @size-border * 2 + @size-indent-small;
-
-.fontSelectContainer {
-    overflow: visible;
-    position: relative;
-}
-
-.FontSelect {
-    border: @size-border solid @color-border;
-    left: @fontSelectCheckboxMargin;
-    margin: 0 !important;
-    overflow: hidden !important; // Hack: Chrome 62 fix.
-    position: absolute;
-    top: 0;
-    width: @size-content-width - @fontSelectCheckboxMargin;
-    z-index: 3;
-
-    .fontSelectLine {
-
-        .TextBox.fontSelectTextBox {
-            border: none;
-
-            &:not(:first-child) {
-                border-top: @size-border/2 solid @color-border;
-            }
-        }
-
-        .Button.fontSelectTick {
-            border: none;
-
-            .icon.down {
-                position: relative;
-
-                &::before,
-                &::after {
-                    background-color: @color-input-fore-active;
-                    content: "";
-                    display: inline-block;
-                    height: 0.25em;
-                    position: absolute;
-                    top: 0.875em;
-                    width: 0.75em;
-                }
-
-                &::before {
-                    left: 0.25em;
-                    transform: skewY(45deg);
-                }
-
-                &::after {
-                    left: 1em;
-                    transform: skewY(-45deg);
-                }
-            }
-        }
-        
-        &::after {
-            opacity: 0;
-            transition: opacity @time-slow;
-        }
-    }
-
-    .fontList {
-        background-color: @color-back;
-        max-height: 16em;
-        transition: max-height @time-slow*2;
-
-        &.collapsed {
-            max-height: 0;
-            transition: max-height @time-fast;
-        }
-
-        .fontItem {
-            background-color: @color-back;
-            color: @color-description;
-            cursor: pointer;
-            justify-content: flex-start;
-            padding: 0 @size-indent-small;
-
-            &:hover {
-                background-color: @color-hover;
-            }
-
-            &:active {
-                background-color: @color-active;
-                color: @color-input-fore-active;
-            }
-        }
-    }
-    
-    &.renderingFonts {
-        .fontSelectLine::after {
-            color: @color-input-fore-placeholder;
-            content: "Loading...";
-            display: inline-block;
-            line-height: @size-control-inner;
-            opacity: 1;
-            position: absolute;
-            right: @size-control-inner+@size-indent-small;
-            top: 0;
-        }
-    }
-}
