--- src/popup/style/_site_list.less
@@ -1,20 +1,0 @@
-﻿@import "_variables.less";
-
-/*---------------
-    Site list
------------------*/
-
-.SiteList {
-    border: @size-border solid @color-border;
-    flex: none;
-    overflow: hidden !important; // Hack: Chrome 62 fix.
-
-    .TextBox.siteListTextBox {
-        border: none;
-        height: 1.5em;
-
-        &:not(:first-child) {
-            border-top: @size-border/2 solid @color-border;
-        }
-    }
-}
