--- src/popup/style/_tab_panel.less
@@ -1,107 +1,0 @@
-﻿@import "_variables.less";
-
-/*-------------
-    Tab panel
----------------*/
-
-.TabPanel {
-    
-    .tabPanelButtons {
-        align-items: stretch;
-        flex: none;
-        display: flex;
-        flex-flow: row nowrap;
-        justify-content: center;
-        // WARNING: Some flexbox bug with border-top at child.
-        // A scrollbar appears.
-        overflow-y: hidden;
-        width: 100%;
-
-        .tabPanelButton {
-            //background-color: transparent;
-            border-bottom: @size-border solid @color-border;
-            border-left: 0 solid @color-border;
-            border-right: 0 solid @color-border;
-            border-top: 0 solid @color-border;
-            cursor: pointer;
-            display: inline-block;
-            height: @size-control-inner*1.1;
-            line-height: @size-control-inner*1.1;
-            font-weight: bold;
-            //border-top: @size-border solid transparent;
-            padding-bottom: 0;
-            padding-left: @size-indent-small+@size-border;
-            padding-right: @size-indent-small+@size-border;
-            padding-top: 0;
-            //transition: background-color @time-slow,border-width @time-slow,padding @time-slow;
-            transition: background-color @time-slow;
-
-            .text {
-                color: @color-border;
-                font-size: 1.2em;
-            }
-
-            &.active {
-                border-top: @size-border solid @color-border;
-                border-right: @size-border solid @color-border;
-                //border-bottom: @size-border solid transparent;
-                border-bottom: 0 solid @color-border;
-                border-left: @size-border solid @color-border;
-                cursor: default;
-                padding-left: @size-indent-small;
-                padding-right: @size-indent-small;
-
-                .text {
-                    color: @color-header;
-                }
-            }
-
-            &:hover:not(.active) {
-                background-color: @color-hover;
-                transition: background-color @time-fast;
-
-                .text {
-                    color: @color-description;
-                }
-            }
-        }
-
-        &::before,
-        &::after {
-            border-bottom: @size-border solid @color-border;
-            content: "";
-            display: inline-block;
-            flex: 1;
-        }
-    }
-
-    .tabPanelContainer {
-        //height: 21em;
-        flex: 1 1 auto;
-        margin-top: @size-indent-big;
-        overflow-y: auto;
-
-        .Tab {
-            overflow-x: hidden;
-            transition: width @time-slow, opacity @time-slow;
-            width: @size-content-width;
-
-            &:not(.active) {
-                //display: none;
-                opacity: 0;
-                overflow: hidden;
-                pointer-events: none;
-                transition: width @time-slow, opacity @time-fast;
-                width: 0;
-            }
-
-            > *:not(:first-child) {
-                margin-top: @size-indent-small;
-            }
-        }
-    }
-    
-    &.disabled {
-        opacity: 0.5;
-    }
-}
