--- src/popup/style/_toggle.less
@@ -1,46 +1,0 @@
-﻿@import "_variables.less";
-
-/*-----------------
-   Toggle control
--------------------*/
-
-.Toggle {
-    border: @size-border solid @color-border;
-    box-sizing: content-box;
-    display: flex;
-    position: relative;
-    user-select: none;
-
-    .on, .off {
-        color: @color-fore;
-        cursor: pointer;
-        display: inline-block;
-        height: @size-control-inner;
-        line-height: @size-control-inner;
-        text-align: center;
-        transition: background-color @time-slow;
-        white-space: nowrap;
-        width: 50%;
-        z-index: 1;
-
-        &:hover:not(:active):not(.active) {
-            background-color: @color-hover;
-            transition: background-color @time-fast;
-        }
-    }
-    
-    // Slider
-    &::before {
-        background-color: @color-active;
-        content: "";
-        height: 100%;
-        left: 50%;
-        position: absolute;
-        transition: left @time-fast;
-        width: 50%;
-    }
-    
-    &.switchedOn::before{
-        left: 0;
-    }
-}
