--- src/popup/style/_up_down.less
@@ -1,103 +1,0 @@
-﻿@import "_variables.less";
-
-/*-------------------
-    Up-down control
----------------------*/
-
-.UpDown {
-    .controlContainer();
-
-    .buttonDown,
-    .buttonUp {
-        position: relative;
-
-        &::before,
-        &::after {
-            background-color: @color-fore;
-            content: "";
-            display: inline-block;
-            position: absolute;
-        }
-    }
-
-    .buttonDown {
-
-        &::before {
-            height: 0.75em;
-            left: 0.8125em;
-            top: 0.25em;
-            transform: skewX(-45deg);
-            width: 0.25em;
-        }
-
-        &::after {
-            height: 0.75em;
-            left: 0.8125em;
-            top: 1em;
-            transform: skewX(45deg);
-            width: 0.25em;
-        }
-    }
-
-    .buttonUp {
-
-        &::before {
-            height: 0.75em;
-            left: 1em;
-            top: 0.25em;
-            transform: skewX(45deg);
-            width: 0.25em;
-        }
-
-        &::after {
-            height: 0.75em;
-            left: 1em;
-            top: 1em;
-            transform: skewX(-45deg);
-            width: 0.25em;
-        }
-    }
-    // Do not change opacity, but icon color when disabled
-    .Button.disabled {
-        opacity: 1;
-
-        .buttonDown,
-        .buttonUp {
-
-            &::before,
-            &::after {
-                background-color: @color-input-fore-placeholder;
-            }
-        }
-    }
-}
-
-.TrackBar {
-    border: @size-border solid @color-border;
-    overflow: hidden;
-    position: relative;
-    user-select: none;
-    width: 100%;
-
-    .value {
-        background-color: @color-active;
-        display: inline-block;
-        height: @size-control-inner;
-        transition: width @time-slow;
-        width: 30%;
-    }
-
-    label {
-        color: @color-fore;
-        display: inline-block;
-        height: @size-control-inner;
-        left: 0;
-        line-height: @size-control-inner;
-        position: absolute;
-        text-align: center;
-        top: 0;
-        white-space: nowrap;
-        width: 100%;
-        z-index: 2;
-    }
-}
