--- src/ui/controls/checkbox.tsx
@@ -1,15 +1,0 @@
-import { html } from 'malevic';
-import { mergeClass, omitAttrs } from './utils';
-
-export default function CheckBox(props: Malevic.NodeAttrs) {
-    props = props || {};
-    const cls = mergeClass('checkbox', props.class);
-    const attrs = omitAttrs(['class', 'checked', 'onchange'], props);
-
-    return (
-        <label class={cls} {...attrs}>
-            <input class="checkbox__input" type="checkbox" onchange={props.onchange} />
-            <span class="checkbox__checkmark"></span>
-        </label>
-    );
-}
