+++ src/ui/controls/index.ts
@@ -6,7 +6,7 @@ import Select from './select';
 import Shortcut from './shortcut';
 import TabPanel from './tab-panel';
 import TextBox from './textbox';
-import TextList from './text_list';
+import TextList from './text-list';
 import Toggle from './toggle';
 import UpDown from './updown';
