+++ src/ui/controls/shortcut/style.less
@@ -1,0 +1,7 @@
+@import "../../theme";
+
+.shortcut {
+    display: inline-block;
+    font-size: @size-text-small;
+    text-align: center;
+}
