+++ src/ui/controls/style.less
@@ -1,0 +1,7 @@
+@import "./button/style";
+@import "./col/style";
+@import "./row/style";
+@import "./shortcut/style";
+@import "./tab-panel/style";
+@import "./toggle/style";
+@import "./updown/style";
