+++ src/ui/controls/toggle/style.less
@@ -1,0 +1,44 @@
+@import "../../theme";
+
+.toggle {
+    background-color: @color-control-back;
+    border: @size-border solid @color-border;
+    box-sizing: content-box;
+    color: @color-control-fore;
+    display: inline-block;
+    height: @size-control-inner;
+    position: relative;
+
+    &__btn {
+        align-items: center;
+        cursor: pointer;
+        display: inline-flex;
+        justify-content: center;
+        height: 100%;
+        position: relative;
+        transition: background-color @time-slow;
+        width: 50%;
+
+        &:hover:not(&--active) {
+            background-color: @color-control-hover;
+        }
+    }
+
+    &::before {
+        background-color: @color-control-active;
+        content: "";
+        display: inline-block;
+        height: 100%;
+        left: 50%;
+        position: absolute;
+        top: 0;
+        transition: left @time-fast;
+        width: 50%;
+    }
+
+    &--checked {
+        &::before {
+            left: 0;
+        }
+    }
+}
