+++ src/ui/controls/utils.ts
@@ -1,12 +1,12 @@
-import { classes } from 'malevic';
+import {classes} from 'malevic';
 
 function toArray<T>(x: T | T[]) {
     return Array.isArray(x) ? x : [x];
 }
 
 export function mergeClass(
-    cls: string | { [cls: string]: any; } | (string | { [cls: string]: any; })[],
-    propsCls: string | { [cls: string]: any; } | (string | { [cls: string]: any; })[]
+    cls: string | {[cls: string]: any;} | (string | {[cls: string]: any;})[],
+    propsCls: string | {[cls: string]: any;} | (string | {[cls: string]: any;})[]
 ) {
     const normalized = toArray(cls).concat(toArray(propsCls));
     return classes(...normalized);
