+++ src/ui/popup/components/body.tsx
@@ -3,8 +3,8 @@ import withForms from 'malevic/forms';
 import {TabPanel, Button, CheckBox} from '../../controls';
 import TopSection from './top-section';
 import FilterSettings from './filter-settings';
-import FontSettings from './font_settings';
-import SiteListSettings from './site_list_settings';
+import FontSettings from './font-settings';
+import SiteListSettings from './site-list-settings';
 import {Extension} from '../../../definitions';
 
 withForms();
