+++ src/ui/popup/components/filter-settings/style.less
@@ -1,0 +1,59 @@
+.filter-settings {
+    align-items: stretch;
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+
+    > :not(:first-child) {
+        margin-top: @indent-large;;
+    }
+}
+
+.mode-toggle {
+    align-items: stretch;
+    display: flex;
+    flex-direction: column;
+
+    &__line {
+        display: flex;
+        flex-direction: row;
+
+        > :not(:first-child) {
+            margin-left: @indent-small;
+        }
+    }
+
+    .toggle {
+        flex: auto;
+    }
+
+    &__label {
+        display: inline-block;
+        font-size: @size-text-small;
+        line-height: @size-text-small;
+        text-align: center;
+    }
+
+    &__button--active {
+        background: @color-control-active;
+
+        &:hover {
+            background: @color-control-active;
+        }
+    }
+}
+
+.icon {
+    background-size: contain;
+    display: inline-block;
+    height: 100%;
+    width: 100%;
+}
+
+.icon--dark-mode {
+    background-image: url(../assets/images/mode-dark-32.svg);
+}
+
+.icon--light-mode {
+    background-image: url(../assets/images/mode-light-32.svg);
+}
