+++ src/ui/popup/components/site-toggle/style.less
@@ -1,0 +1,13 @@
+.site-toggle {
+    color: @color-fore;
+    white-space: pre-wrap;
+
+    &--disabled {
+        opacity: 0.5;
+        pointer-events: none;
+    }
+
+    &__url {
+        color: @color-control-fore;
+    }
+}
