+++ src/ui/popup/components/top-section/style.less
@@ -1,0 +1,21 @@
+.top-section {
+    display: flex;
+    flex: auto;
+    flex-direction: row;
+    width: 100%;
+
+    &__control {
+        display: flex;
+        flex: auto;
+        flex-direction: column;
+
+        &:not(:first-child) {
+            margin-left: @indent-small;
+        }
+    }
+}
+
+#site-toggle-container {
+    flex: none;
+    width: 55%;
+}
