--- src/ui/style/controls/button.less
@@ -1,29 +1,0 @@
-@import "../theme";
-
-.button {
-    align-items: center;
-    background-color: @color-control-back;
-    border: @size-border solid @color-border;
-    box-sizing: content-box;
-    color: @color-control-fore;
-    cursor: pointer;
-    display: inline-flex;
-    height: @size-control-inner;
-    justify-content: center;
-    line-height: @size-text-normal;
-    min-width: @size-control-inner;
-    overflow: hidden;
-    padding: 0;
-    text-align: center;
-    transition: background-color @time-slow;
-    user-select: none;
-
-    &:hover {
-        background-color: @color-control-hover;
-        transition: background-color @time-fast;
-    }
-
-    &:active {
-        background-color: @color-control-active;
-    }
-}
