--- src/ui/style/controls/index.less
@@ -1,7 +1,0 @@
-@import "./button";
-@import "./col";
-@import "./row";
-@import "./shortcut";
-@import "./tab-panel";
-@import "./toggle";
-@import "./updown";
