--- src/ui/style/controls/shortcut.less
@@ -1,7 +1,0 @@
-@import "../theme";
-
-.shortcut {
-    display: inline-block;
-    font-size: @size-text-small;
-    text-align: center;
-}
