--- src/ui/style/controls/tab-panel.less
@@ -1,69 +1,0 @@
-.tab-panel {
-    align-items: stretch;
-    display: flex;
-    flex-direction: column;
-
-    &__button {
-        border-bottom: @size-border solid @color-border;
-        border-left: none;
-        border-right: none;
-        border-top: none;
-        box-sizing: border-box;
-        color: mix(@color-fore, @color-control-back, 50%);
-        font-size: @size-text-large;
-        font-weight: bold;
-        height: @size-control-inner + @size-border;
-        padding: 0 @indent-small + @size-border;
-
-        &:hover {
-            color: @color-fore;
-        }
-
-        &:active {
-            background-color: @color-control-hover;
-        }
-
-        &--active {
-            border-bottom: none;
-            border-left: @size-border solid @color-border;
-            border-right: @size-border solid @color-border;
-            border-top: @size-border solid @color-border;
-            color: @color-heading;
-            padding: 0 @indent-small;
-
-            &:hover {
-                color: @color-heading;
-            }
-        }
-    }
-
-    &__buttons {
-        display: flex;
-        flex: none;
-        flex-direction: row;
-        justify-content: center;
-
-        &::before,
-        &::after {
-            border-bottom: @size-border solid @color-border;
-            content: "";
-            flex: auto;
-        }
-    }
-
-    &__tabs {
-        display: flex;
-        flex: auto;
-        flex-direction: row;
-        margin-top: @indent-large;
-    }
-
-    &__tab {
-        display: none;
-        width: 100%;
-
-        &--active {
-            display: flex;
-        }
-    }
-}
